Create 2D_Molasse_170531
Here, we used the redback branch to model the deformation and stress history in a temperature and strain rate sensitive framework.
It is a layered model with a displacement bc on one side and fixed bc on the other side and at the bottom. To the model there is a temperature applied.
The model set up is a layered model with an integrated fault structure as an inhomogenity. 
Model steps:
t = 0-100:
- Temperature gradient
- gravity is working on the model
t= 100 - end
- tectonic deformation rate of 10^-14 m/s is applied from one side of the model -> so that the stress and deformation in the model is increased.